Rename customer form
