--- packages/ui/src/Forms/CustomerDetails/index.ts
@@ -1,3 +1,0 @@
-import CustomerDetails from "./CustomerDetails";
-
-export default CustomerDetails;
