+++ packages/ui/src/Forms/CustomerProfile/index.ts
@@ -1,0 +1,3 @@
+import CustomerProfileForm from "./CustomerProfileForm";
+
+export default CustomerProfileForm;
